+++ CHANGELOG.md
@@ -4,6 +4,53 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+---
+### 3.5.0 - 2023-01-21
+
+### Update
+
+- rich-domain: update lib core to 1.17.0
+
+
+---
+### 3.4.7 - 2023-01-19
+
+### Update
+
+- rich-domain: update lib core to 1.16.2
+
+---
+### 3.4.6 - 2023-01-18
+
+### Update
+
+- rich-domain: update lib core to 1.16.1
+
+---
+### 3.4.5 - 2023-01-14
+
+### Changed
+
+- date.value-object: rename method from `isEqual` to `isEqualDate`
+
+### Update
+
+- rich-domain: update lib core to 1.16.0
+- Entity: added method isEqual to compare current instance with another one.
+- ValueObject: added method isEqual to compare current instance with another one. [Issue 27](https://github.com/4lessandrodev/rich-domain/issues/27)
+
+---
+
+### 3.4.4 - 2023-01-12
+
+### Added
+
+- custom-string.value-object: By: [VinnyLima](https://github.com/VinnyLima)
+- removeSpecialChars and onlyNumbers: [Issue 223](https://github.com/4lessandrodev/types-ddd/issues/223)
+- email.value-object: added MESSAGE as customizable value
+
+---
+
 ### 3.4.3 - 2023-01-05
 
 ### Updated